requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
openai==1.68.2
faiss-cpu==1.10.0
numpy==2.2.4
tiktoken==0.9.0
python-dotenv==1.0.1
pytest==8.3.5
httpx==0.28.1
